+++ src/py42/services/util.py
@@ -1,3 +1,5 @@
+import re
+
 import py42.settings as settings
 
 
@@ -13,3 +15,18 @@ def get_all_pages(func, key, *args, **kwargs):
         yield response
         page_items = response[key]
         item_count = len(page_items)
+
+
+def escape_quote_chars(token):
+    """
+    The `nextPgToken` returned in Forensic Search requests with > 10k results is the eventId
+    of the last event returned in the response. Some eventIds have double-quote chars in
+    them, which need to be escaped when passing the token in the next search request.
+    """
+    unescaped_quote_pattern = r'[^\\]"'
+
+    return re.sub(
+        pattern=unescaped_quote_pattern,
+        repl=lambda match: match.group().replace('"', r"\""),
+        string=token,
+    )